pyproject.toml
[tool.poetry]
name = "game-server"
version = "0.1.0"
description = "game server"
authors = ["gleb <glebbarko@mail.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
alembic = "^1.15.2"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = "^6.0.0"
aioredis = "^2.0.1"
pydantic = {extras = ["email"], version = "^2.11.4"}
python-dotenv = "^1.1.0"
pydantic-settings = "^2.9.1"
greenlet = "^3.2.1"
bcrypt = "^3.2.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
async-asgi-testclient = "^1.4.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
